==> notebooks/03_tef_plant_pics.Rmd <==
---
title: "Sev Plant Summary"
author: "Tim Farkas"
date: "4/19/2018"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```

```{r Load Libraries}
library(tidyverse)
library(ape)
library(geiger)
```


##### Modify Trait Data For Analysis 
```{r Data Input and Prep}
raw_traits <- read_csv("~/Dropbox/Projects/Sev/Plants/Data/All Traits/sev_all_traits_Jan2020.csv") %>%
  
  # just not sure about this ID, plust not much data
  filter(! kartez %in% c("DAJA?")) %>% 
  
  # lump big Aristida under ARIST
  mutate_at(vars(kartez), funs(ifelse(. == "ARPU9", "ARIST", .))) %>% 
  
  # lump Sporobolus 
  mutate_at(vars(kartez), funs(ifelse(. %in% c("SPCO4", "SPFL2"), "SPORO", .))) %>% 
  
  # lump Astragalus
  mutate_at(vars(kartez), funs(ifelse(. %in% c("ASFE2", "ASNU4"), "ASTRA", .))) %>% 
  
  # numeric annual vs. perennial
  mutate_at(vars(a_p), funs(ifelse(. == "a", 0, 1))) %>% 
  
  # group by plant species
  group_by(kartez) 

load("~/Dropbox/1_Work/1_Research/Whitney-Rudgers Lab/Sev/Plants/sev_traits_cv_manuscript/Data/raw_bounce_wide_q1090m.RData")

all_traits <- 
  # now there are duplicate kartez codes, so summarize
  inner_join(raw_traits %>% summarize_at(vars(family:apgf, -a_p), funs(max)),
             raw_traits %>% summarize_at(vars(a_p, avg.height_A:seed_mass), funs(mean)), 
             by = "kartez") %>% 
  # remove outliers
  mutate_at(vars(lma), funs(ifelse(. > 0.1, NA, .))) %>%
  mutate_at(vars(p95.height_A), funs(ifelse(. > 100, NA, .))) %>%
  mutate_at(vars(sdens), funs(ifelse(. > 0.003, NA, .))) %>%
  mutate_at(vars(rdens), funs(ifelse(. > 0.003, NA, .))) %>%
  mutate_at(vars(fr.diam), funs(ifelse(. > 0.15, NA, .))) %>%

# join in cv values
  inner_join(temp_wide)
```

#### PICs

```{r Single Trait Pics Function With Node Matching}

pics <- function(x, # a tibble or data frame with trait data
                 trait, # a character string of a single trait column name
                 phy, # a phylogenetic tree of class "phylo" & order "cladewise"
                 taxa, # a character string of the taxon column, to match tips in phy
                 scale=TRUE) { # logical indicator whether to mean center PIC output
  
  require(ape)
  require(dplyr)

  #### modify data ####
  
    dd <- x %>%
      as_tibble() %>%
      select(taxa, trait) %>%
      filter(complete.cases(.)) %>% # filter complete cases
      as.data.frame() # to data frame in case of tibble
    # get row names for ape::pic function
    taxon.names <- dd[, taxa]
    
    # get trait values data frame
    traits <- dd[, trait, drop=FALSE]
    
    # assign row names
    rownames(traits) <- taxon.names
  
  #### modify tree ####   
    
    # resolve multichotomies
    new_phy <-  multi2di(phy, random=FALSE)
    
    # label nodes
    new_phy$node.label <- paste("node", 
                            formatC(c(1:new_phy$Nnode), 
                                    digits=0, width=nchar(new_phy$Nnode), flag="0"))
    
    # prune tree and data
    pruned <- drop.tip(new_phy, setdiff(phy$tip.label, taxon.names))
     
    traits <- traits[! rownames(traits) %in% setdiff(phy$tip.label, taxon.names),
                     , drop=FALSE]

    

    
  #### apply pic function ####
    
    pic.out <- pic(x = set_names(traits[[1]], rownames(traits)), 
                   phy = pruned, 
                   scaled = scale, 
                   var.contrasts = TRUE)
    pic.out <- pic.out %>%
      as_tibble() %>%
      mutate(node = rownames(pic.out)) %>%
      mutate(pic_abs = abs(contrasts), 
             sd = sqrt(variance)) %>%
      select(node, pic=contrasts, pic_abs, var=variance, sd)
    
    return(pic.out)
    #return(pruned)
}

```

```{r Find Transformation}

brl_transform <- function(x, # a tibble or data frame with trait data
                          trait, # a character string of trait column
                          phy, # a phylogenetic tree of class "phylo" & order "cladewise"
                          taxa, # a character string of the taxon column, to match tips in phy
                          scale=TRUE) {
  
  require(geiger)
  require(tidyverse)
  
  phy.all1 <- compute.brlen(phy=phy, runif, min=0, max=1) # all equals one is base
  
  # raw
  pic.raw <- list(pics(x=x, trait=trait, phy=phy, taxa=taxa, scale=scale) %>% 
                    mutate(trans = "raw"))
  
  # All equals 1
  pic.all1 <- list(pics(x=x, trait=trait, phy=phy.all1, taxa=taxa, scale=scale) %>% 
                    mutate(trans = "all1"))
  

  # Pagel transformations
  lambdas <- seq(0, .9, by=.1)  
  
  pic.pagel <- lapply(lambdas, function(l) {
    
    phy.pagel <- rescale(x=phy.all1, model="lambda", l)
    pics(x=x, trait=trait, phy=phy.pagel, taxa=taxa, scale=scale) %>%
      as_tibble() %>%
      mutate(trans = paste("Pagel: lambda = ", l, sep=""))
    
  }) 
  
  names(pic.pagel) <- paste("Pagel: lambda = ", lambdas, sep="")
  
  # Grafen "transformations"
  rhos <- seq(.1, 1, by=.1)
  
  pic.grafen <- lapply(rhos, function(p) {
    
    phy.grafen <- compute.brlen(phy=phy.all1, power=p)
    pics(x=x, trait=trait, phy=phy.grafen, taxa=taxa, scale=scale) %>%
      as_tibble() %>%
      mutate(trans = paste("Grafen: power = ", p, sep=""))
    
  }) 
  
  names(pic.grafen) <- paste("Grafen: power = ", rhos, sep="")
  
  # Ornstein-Uhlenbeck transformations
  alphas <- c(.01, .05, .1, .25, .5, .75, 1, 2)
  
  pic.OU <- lapply(alphas, function(alp) {
    
    phy.OU <- rescale(x=phy.all1, model="OU", alp)
    pics(x=x, trait=trait, phy=phy.OU, taxa=taxa, scale=scale) %>%
      as_tibble() %>%
      mutate(trans = paste("OU: alpha = ", alp, sep=""))
    
  }) 
  
    names(pic.OU) <- paste("OU: alpha = ", alphas, sep="")
    
  # combine into one list
  pics.out <- c(raw=pic.raw, all1=pic.all1, 
                pic.pagel, pic.grafen, pic.OU)
  
  return(pics.out)
  
}


```

```{r Function Testing}
x <- all_traits 
trait <- "sdens"
taxa="taxon.new"
phy <- read.tree("Data/sev_traits_tree101.tre")

# test single pic function
test_pic <- pics(x=all_traits, trait=trait, phy=read.tree("Data/sev_traits_tree101.tre"), taxa="taxon.new")

# test branch length transformation function
pics.ldmc <- brl_transform(x=all_traits, trait=trait, phy=read.tree("Data/sev_traits_tree101.tre"), taxa = "taxon.new")

```

```{r Visualize Transformations}

# get all trait names
trait_names <- all_traits %>%
  select(a_p:F_P_BM_cv) %>%
  names()

# just variance metrics
trait_names <- all_traits %>%
  select(F_A_BM_acor:F_P_BM_cv) %>%
  names()


# loop through names to create diagnostic transformation plots
for(trait in trait_names) {
  
  print(trait)
  
  # get pics for all transformations
  trait_pics <- brl_transform(x=all_traits, trait=trait, 
                              taxa = "taxon.new",
                              phy=read.tree("~/Dropbox/Projects/Sev/Plants/sev_traits_cv_manuscript/Data/sev_traits_tree101.tre"))
  
  # bind to single data frame
  trait.pics.df <- do.call(rbind, trait_pics) 
    # filter(!grepl("power", .$trans)) %>% # Pagel only
   
  # get correlations
  trait.cors <- trait.pics.df %>%
    group_by(trans) %>%
    summarize(cor=round(cor(pic_abs, sd), 2))
  
  # create diagnostic plot

  trait.pics.df %>%
    ggplot(aes(x=pic_abs, y=sd)) +
    geom_label(aes(label=substr(node, nchar(node) - 2, nchar(node))),
               cex=2.5) +
    facet_wrap(~ trans, scales = "free") + 
    geom_text(data = trait.cors, 
              aes(x=Inf, y=Inf, label=paste("r = ", trait.cors$cor, sep=""),
                  vjust=1, hjust=1),
              cex = 3, col="red")
  
  # write to disk
  ggsave(filename = paste("~/Dropbox/Projects/Sev/Plants/sev_traits_cv_manuscript/Figures/pic_trans_diagnostics_", trait, ".pdf", 
                          sep=""),
         device="pdf", height = 14, width=14)
  
}

```

```{r Chose Transformations}


# manually chosen transformations
trans <- c(a_p=3, area=3, avg.height_A=3, avg.height_B=3, avg.height_C=2, 
           avg.height_F=2, avg.height_G=28, avg.height_P=2, cn=2, d13C=3, 
           d15N=2, F_A_BM_cv=2, F_B_BM_cv=28, F_C_BM_cv=24, F_F_BM_cv=2,
           F_G_BM_cv=2, F_P_BM_cv=21, fr.dens= 3, fr.diam=3, fr.dmc=3,
           fr.srl=3, ldmc=2, lma=4, max.height_A=3, max.height_B=3, 
           max.height_C=2, max.height_F=2, max.height_G=3, max.height_P=2,
           p95.height_A=2, p95.height_B=2, p95.height_C=2, p95.height_F=2, 
           p95.height_G=4, p95.height_P=2, pC=2, pN=2, rdens=2, rdmc=28, 
           sdens=2, sdmc=2, seed_mass=3, sla=2, 
           F_A_BM_q1090=2, F_B_BM_q1090=29, F_C_BM_q1090=29, 
           F_F_BM_q1090=2, F_G_BM_q1090=2, F_P_BM_q1090=2)

# ugg, some faff to associate correct transformation with choice from 
# output figures
names <- names(all_traits)[15:63]
names(trans) <- names[order(names)]

trans_options <- names(brl_transform(x=all_traits, trait="a_p",
               read.tree("~/Dropbox/Projects/Sev/Plants/sev_traits_cv_manuscript/Data/sev_traits_tree101.tre"), 
               taxa = "taxon.new"))
trans_options <- trans_options[order(trans_options)]

# perform all transformations for traits of interest, then select
# manually chosen option and combine into one long list
all_pics_list <- setNames(lapply(names(trans), function(t) {
  
 all_trans <- brl_transform(x=all_traits, trait=t,
               read.tree("~/Dropbox/Projects/Sev/Plants/sev_traits_cv_manuscript/Data/sev_traits_tree101.tre"), 
               taxa = "taxon.new")
 all_trans[[trans_options[trans[t]]]]
 
 }
 
 ), names(trans))

# join all traits together

all_pics_tib <- lapply(names(all_pics_list), function(t) {
  
  all_pics_list[[t]] %>%
    select(node, 
           !!t:= pic)
}) %>%
  reduce(full_join, by="node")

write_csv(all_pics_tib, "~/Dropbox/Projects/Sev/Plants/sev_traits_cv_manuscript/Data/all_pics_joined_1Jan2021.csv")



```

## Drop rare species
```{r}
# get ranked list of zeros
temp_cv <- read_csv(paste0(repo_home, "data/interim/reponse_vars_spp_tidy.csv"))

zranks <- temp_cv %>%
  filter(site == "F", 
         allometry == "BM", 
         season == "F", 
         metric == "cv") %>%
  mutate(zrank = rank(14-n_zero_mean)) %>%
  left_join(eval(.) %>%
              group_by(zrank) %>%
              summarize(nsp=n()) %>%
  mutate(nexclude1 = cumsum(nsp), 
         nexclude = lag(nexclude1)) %>%
  mutate(across(nexclude, .f = ~ ifelse(is.na(.x), 0, .x))) %>%
  select(zrank, nexclude), 
by = "zrank") %>%
  arrange(nexclude) %>%
select(kartez, nexclude) 

# join in rank to all traits data
all_traits2 <- all_traits %>%
  left_join(zranks) %>%
  select(kartez, nexclude, family:last_col()) %>%
  drop_na(nexclude)

ranks <- unique(zranks$nexclude)[order(unique(zranks$nexclude))]
tree <- read.tree("~/Dropbox/Projects/Sev/Plants/sev_traits_cv_manuscript/Data/sev_traits_tree101.tre")

inranks <- ranks[1:20]
# loop through species rareness
rare_list <- lapply(inranks, function(rank) {

all_pics_list <- setNames(lapply(names(trans), function(t) {
  
 all_trans <- brl_transform(x=all_traits2 %>% filter(nexclude >= rank), 
                            trait=t, phy=spp_tree,
               taxa = "taxon.new")
 all_trans[[trans_options[trans[t]]]]
 
 }
 
 ), names(trans))

# join all traits together

all_pics_tib <- lapply(names(all_pics_list), function(t) {
  
  all_pics_list[[t]] %>%
    select(node, 
           !!t:= pic)
}) %>%
  reduce(full_join, by="node")

return(all_pics_tib)

}) %>% setNames(inranks)

save(rare_list, 
     file= "~/Dropbox/Projects/Sev/Plants/sev_traits_cv_manuscript/Data/rare_list.RData")

# loop through z_nero mean rank, do analysis, and plot
walk(.x = as.character(inranks), .f = ~ {
  
print(.x)
  
pd <- rare_list[[.x]] %>%
mutate_at(vars(node), function(x) as.character(as.numeric(word(x, 2, 2)))) %>%
  mutate(index = -1 * (a_p + cn + fr.diam + ldmc + lma + rdmc + sdmc + seed_mass)) %>%
  mutate(avc = a_p - cn )

pd_long <- pd %>%
  gather(key = "trait", value = "pic", a_p:d15N, fr.dens:index)

f.list <- names(pd)[c(2:3, 10, 25:30, 37, 45:49)]

## regression

pees <- c(lapply(setNames(f.list, f.list), function(x) {
  
 summary(lm(F_F_BM_q1090 ~ pic, data=pd_long %>% filter(trait==x)))$coefficients[2, 4]
  
}), recursive=TRUE)

sigs <- ifelse(pees > 0.10, "", 
               ifelse(pees > 0.05, "#",
                      ifelse(pees > 0.01, "*", 
                             ifelse(pees > 0.001, "**", "***"))))

reg_out <- tibble(trait=names(pees), p=pees, sig=sigs)

# graph traits vs q1090

pd_long %>%  
  filter(trait %in% f.list) %>%
  ggplot(mapping=aes(y=F_F_BM_q1090, x=pic)) +
  geom_point() + 
    facet_wrap(~ trait, scales="free") +
  geom_text(data = reg_out, 
              aes(x=Inf, y=Inf, label=reg_out$sig,
                  vjust=1, hjust=1),
              cex = 5, col="red") +
  geom_label(aes(label=node), cex=2) +
  geom_smooth(method="loess") +
  ggtitle(paste("rank = ", .x))

ggsave(paste("~/Dropbox/Projects/Sev/Plants/sev_traits_cv_manuscript/Figures/nzero_diagnostics/q1090_vs_traits/q1010_reg_flats_nzero_", .x, ".pdf", sep=""), height=14, width = 14)

# graph traits vs cv
pd_long %>%  
  filter(trait %in% f.list) %>%
  ggplot(mapping=aes(y=F_F_BM_cv, x=pic)) +
  geom_point() + 
    facet_wrap(~ trait, scales="free") +
  geom_text(data = reg_out, 
              aes(x=Inf, y=Inf, label=reg_out$sig,
                  vjust=1, hjust=1),
              cex = 5, col="red") +
  geom_label(aes(label=node), cex=2) +
  geom_smooth(method="loess") +
  ggtitle(paste("rank = ", .x))

ggsave(paste("~/Dropbox/Projects/Sev/Plants/sev_traits_cv_manuscript/Figures/nzero_diagnostics/cv_vs_traits/cv_reg_flats_nzero_", .x, ".pdf", sep=""), height=14, width = 14)

# graph cv vs q1090
pd_long %>%  
  filter(trait %in% f.list) %>%
  ggplot(mapping=aes(y=F_F_BM_cv, x=F_F_BM_q1090)) +
  geom_point() + 
  geom_smooth(method="loess") +
  ggtitle(paste("rank = ", .x))

ggsave(paste("~/Dropbox/Projects/Sev/Plants/sev_traits_cv_manuscript/Figures/nzero_diagnostics/q1090_vs_cv/q1090_vs_cv_flats_nzero_", .x, ".pdf", sep=""), height=7, width = 7)

}
)

```

### Significance tables drop rare species
```{r}
# 1. loop over linear, quadratic, and cubit fits
##  2. loop over rarity-based species sets
###    3. loop over traits
####    4. fit model 
###    5. make traits list of models
##  4. make species-set list of models
# 5. make fit list of models
# 6. loop over fit list (n = 3)
##  7. recursively unlist (fully flatten)
##  8. extract p-value and r-square from each model 
##  9. ggarrange traits as cols, species sets as rows

## alternative
# 1. get list of data frames containing rarity-based species sets
# 2. for each species set:
##  3. 

load("~/Dropbox/Projects/Sev/Plants/sev_traits_cv_manuscript/Data/rare_list.RData")

# list of trait indices
f.list <- names(pd)[c(2:3, 10, 25:30, 37, 45:49)]

# loop to get list of models
mod_list <- 
map(c("linear", "quadratic", "cubic"), .f = function(fit) {
  
  map(as.character(inranks[c(1, 2, 5, 10, 15, 20)]), function(set) {
    
    pd_long <- rare_list[[set]] %>%
    mutate_at(vars(node), 
              function(x) as.character(as.numeric(word(x, 2, 2)))) %>%
    #mutate(index = -1 * (a_p + cn + fr.diam + ldmc + 
    #                       lma + rdmc + sdmc + seed_mass))# %>%
    #mutate(avc = a_p - cn) %>%
    gather(key = "trait", value = "pic", a_p:d15N, fr.dens:sla)
    
    map(c("cv", "q1090"), function(metric) {
      
      map(f.list, function(trt) {
        
        paste(trt, metric, fit)
        form_str <- switch(as.character(fit), 
                           "linear" = paste0("F_F_BM_", metric,
                                        " ~ pic"), 
                           "quadratic" = paste0("F_F_BM_", metric, 
                                        " ~ pic + I(pic^2)"), 
                           "cubic" = paste0("F_F_BM_", metric, 
                                        " ~ pic + I(pic^2) + I(pic^3)"))
  
        mod <- lm(formula(form_str), 
            data=pd_long %>% 
              filter(trait==trt))
        r2 <- summary(mod)$adj.r.squared
        return(list(model = mod, data = mod$model,
                    formula = formula(switch(as.character(fit),
                                             "linear" = "y ~ x",
                                             "quadratic" = "y ~ x + I(x^2)",
                                             "cubic" = "y ~ x + I(x^2) + I(x^3)")),
                    adj.r2 = summary(mod)$adj.r.squared,
                    trait = trt, metric = metric,
                    set = set, fit = fit))
        #return(r2)
        
      }) %>% set_names(f.list)
    }) %>% set_names(c("cv", "q1090"))
  }) %>% set_names(as.character(inranks[c(1, 2, 5, 10, 15, 20)]))
}) %>% set_names(c("linear", "quadratic", "cubic"))

# 
 modtest <- mod_list %>%
  map(~ .x %>%
        map(~ .x %>% transpose)) %>%
  map(~ .x %>% transpose) %>%
  transpose %>%
  map(~ .x %>% transpose) %>%
  map(~ .x %>%
        map(~.x %>% transpose)) %>%
  map(~.x %>%
  unlist(recursive = FALSE)) %>%
   map(~.x %>% unlist(r = FALSE))
   
 
 ggs <- 
  modtest %>%
     map(~ .x %>%
        map(~ {
          gg2 <- .x$data %>%
              rename(x = 1, y = 2) %>%
              ggplot(aes(x, y)) + 
              geom_point(size = 5) + 
              stat_smooth(method = "lm", formula = .x$formula) + 
              xlab(names(.x$data)[1]) + 
              ylab(.x$trait) + 
              ggtitle(paste0("n_spp: ", nrow(.x$data), 
                             "; R-sq: ", round(.x$adj.r2, 2))) + 
              theme(text = element_text(size = 30))  
          gg_list <- list(gg = gg2, meta = .x$trait)
         return(gg_list) 
        }
            )
      ) %>%
   map(~ .x %>% transpose)
 
 ggs %>% 
  walk(~ {
    ga <- .x %>%
      grid.arrange(grobs = .$gg,
                   top=text_grob(.$meta[[1]], size = 50),
                   ncol = 6)
    ggsave(plot = ga, 
           filename = paste0("~/Dropbox/Projects/",
                                        "Sev/Plants/sev_traits_cv_manuscript/",
                                        "Figures/nzero_diognostics_wstats_trait/",
                                        .x$meta[[1]], ".pdf"), 
           width = 40, height = 40, limitsize = FALSE)
  }
  )
  
### something not right with species removal. too many nodes. 
### old stuff

modtest <- map(mod_list, ~ .x %>% transpose) %>% 
  #transpose %>% 
  unlist(recursive = FALSE)

ggs <- 
  modtest %>%
  map(~ .x %>%
  map(~ .x %>%
        map(~ {
          gg2 <- .x$data %>%
              rename(x = 1, y = 2) %>%
              ggplot(aes(x, y)) + 
              geom_point() + 
              stat_smooth(method = "lm", formula = .x$formula) + 
              xlab(names(.x$data)[1]) + 
              ylab(.x$trait) + 
              ggtitle(paste0("rm: ", .x$set, 
                             "; R-sq: ", round(.x$adj.r2, 2)))
          gg_list <- list(gg = gg2, meta = paste(.x$metric, .x$fit, sep = "; "))
         return(gg_list) 
        }
            )
      )
  )

ggs %>% 
  map(~ .x %>%
        unlist(recursive = FALSE) %>%
        transpose() 
      ) %>%
  walk(~ {
    ga <- .x %>%
      grid.arrange(grobs = .$gg, top = .$meta[[1]], ncol = length(f.list))
    ggsave(plot = ga, 
           filename = paste0("~/Dropbox/Projects/",
                                        "Sev/Plants/sev_traits_cv_manuscript/",
                                        "Figures/nzero_diagnostics_wstats/",
                                        .x$meta[[1]], ".pdf"), 
           width = 45, height = 20)
  }
  )

````
